Fix and use Int as the return type of Delete and Update.

+++ app/src/main/java/com/example/personal_financial_manager/data/ExpensesRepository.kt
@@ -5,7 +5,7 @@ import com.example.personal_financial_manager.data.local.entity.Expense
 interface ExpensesRepository {
     suspend fun fetchExpenses(): List<Expense>
     suspend fun insertExpense(expense: Expense): Long
-    suspend fun updateExpense(expense: Expense): Long
-    suspend fun deleteExpense(expense: Expense): Long
-    suspend fun deleteAllExpense(): Long
+    suspend fun updateExpense(expense: Expense): Int
+    suspend fun deleteExpense(expense: Expense): Int
+    suspend fun deleteAllExpense(): Int
 }
